requirements.txt
Faker==11.3.0
kafka-python==2.0.2
py4j==0.10.9.2
pyspark==3.1.1
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
text-unidecode==1.3
typing-extensions==4.0.1
